Suti paranneltu.

+++ shaders/cube_wire.vert
@@ -12,9 +12,9 @@ uniform int z_offset;
 void main()
 {
 
-  float x = mod(index,(x_offset*4)); 
-  float y = mod(index / (x_offset*4),(y_offset*4)); 
-  float z = mod(index / (x_offset * y_offset * 16), z_offset*4); 
+  float x = int(mod(index,(x_offset*4))); 
+  float y = int(mod(index / (x_offset*4),(y_offset*4))); 
+  float z = int(mod(index / (x_offset * y_offset * 16), z_offset*4)); 
 
   gl_Position =  vec4(x,y,z,0.0);
 }
